Create updating_utils module
Moved extract_epub_date to updating_utils
